Fix api base path, add tests
