+++ packages/api/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testPathIgnorePatterns: ['dist'],
+  preset: "@shelf/jest-mongodb",
+};
