+++ packages/api/package.json
@@ -23,7 +23,7 @@
     "build:js": "babel src --out-dir dist --extensions \".ts\" --source-maps inline --copy-files",
     "start": "node ./dist/",
     "clean": "rimraf ./dist",
-    "test": "echo \"jest\"",
+    "test": "jest",
     "lint": "eslint ./ --ext ts",
     "release": "yarn version --no-git-tag-version"
   },
@@ -36,17 +36,20 @@
     "@graphql-codegen/cli": "^2.4.0",
     "@graphql-codegen/typescript": "^2.4.2",
     "@graphql-codegen/typescript-resolvers": "^2.4.3",
+    "@shelf/jest-mongodb": "^2.2.1",
     "@types/graphql-iso-date": "^3.4.0",
     "@types/lodash": "^4.14.161",
     "@types/mongodb": "^3.5.32",
     "@types/node": "^14.11.2",
     "@types/serve-static": "^1.13.10",
+    "@types/supertest": "^2.0.12",
     "@types/uuid": "^8.3.0",
     "babel-plugin-module-resolver": "^4.0.0",
     "jest": "^27.4.7",
     "nodemon": "^2.0.5",
     "npm-run-all": "^4.1.5",
     "pretty-quick": "^2.0.1",
+    "supertest": "^6.3.3",
     "typescript": "^4.0.3"
   },
   "dependencies": {
@@ -66,25 +69,5 @@
     "source-map-support": "^0.5.20",
     "stoppable": "^1.1.0",
     "uuid": "^3.3.2"
-  },
-  "jest": {
-    "testPathIgnorePatterns": [
-      "dist"
-    ],
-    "testMatch": [
-      "<rootDir>/**/__tests__/?(*.)(spec|test).(t)s?(x)"
-    ],
-    "transformIgnorePatterns": [
-      "[/\\\\](node_modules)[/\\\\].+\\.(js|mjs|ts)$"
-    ],
-    "moduleNameMapper": {
-      "^@sorry-cypress/api/(.*)": "<rootDir>/$1"
-    },
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "d.ts"
-    ]
   }
 }
